% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izipr-package.R
\docType{data}
\name{appleshoots}
\alias{appleshoots}
\title{appleshoots data set}
\format{
A data frame with 270 observations on 4 variables.
\describe{
\item{roots}{number of roots per shoot}
\item{trtn}{number of shoots per treatment combination}
\item{photo}{photoperiod, 8 or 16}
\item{bap}{BAP concentration, numeric}
}
}
\source{
The \code{agridat} package
}
\usage{
data(appleshoots)
}
\description{
This data set gives the root counts for propagated columnar apple shootrs, in which the shoots were cultured on different concentrations of the cytokinin BAP, in growth cabinets with an 8 or 16 hour photoperiod. The \code{appleshoots} data has 270 observations.
}
\examples{
## For examples see example(glm.izip)
}
\references{
Wright, K. (2021). agridat: Agricultural Datasets. R package version 1.18. https://CRAN.R-project.org/package=agridat

Ridout, M. S., Hinde, J. P., and Demetrio, C. G. B. (1998). Models for Count Data with Many Zeros. Proceedings of the 19th International Biometric Conference, 179-192.

SAS. Fitting Zero-Inflated Count Data Models by Using PROC GENMOD. support.sas.com/rnd/app/examples/stat/GENMODZIP/roots.pdf
}
\keyword{datasets}
