% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izipr-package.R
\docType{data}
\name{arson}
\alias{arson}
\title{arson data set}
\format{
A vector of 144 observations.
}
\source{
https://www.icpsr.umich.edu/icpsrweb/NACJD
}
\usage{
data(arson)
}
\description{
This data set gives the monthly counts of arson in the 13th police car beat plus in Pittsburgh PA, USA, starting from January 1990 to December 2001. We have 144 observations in which 61 of them, or 42.4\%, are zeros.
}
\examples{
## For examples see example(tsglm.izip)
}
\references{
Zhu, F. (2012) Zero-inflated poisson and negative binomial integer-valued garch models. Journal of Statistical Planning and Inference, 142, 826–839.
}
\keyword{datasets}
