% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{coef.izip}
\alias{coef.izip}
\title{Extract Model Coefficients from an izip Model Fit}
\usage{
\method{coef}{izip}(object, ...)
}
\arguments{
\item{object}{an object class 'izip' object, obtained from a call to \code{glm.izip}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
Coefficients extracted from the object \code{object}.
}
\description{
An function used to extract model coefficients from a 'izip' object.
\code{coefficients} is an alias for \code{coef}.
}
\seealso{
\link{fitted.izip}, \link{residuals.izip}, \link{glm.izip}.
}
