% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izipr-package.R
\docType{package}
\name{compak-package}
\alias{compak-package}
\alias{izipr}
\title{Interpretable Zero-inflated Poisson Regression}
\description{
Interpretable Zero-inflated Poisson Regression
}
\references{
Huang, A., and Fung, T. (2021). \code{izipr}:
Interpretable Zero-inflated Poisson Regression. R package version 0.1

Huang, A., and Fung, T. (2021). A consistent second-order discrete kernel smoother.
Under revision.
}
\keyword{package}
