% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP.R
\name{glm.izip}
\alias{glm.izip}
\title{Fit an interpretable Zero-Inflated Poisson Generalized Linear Model}
\usage{
glm.izip(
  formula,
  data,
  ref.lambda = NULL,
  offset = NULL,
  subset,
  contrasts = NULL,
  na.action
)
}
\arguments{
\item{formula}{an object of class 'formula': a symbolic description of the model to be fitted to the mean via log-link.}

\item{data}{an optional data frame containing the variables in the model}

\item{ref.lambda}{the rate of a Poisson distribution that baseline zero-inflated odds based on.}

\item{offset}{this can be used to specify an \emph{a priori} known component to be included}

\item{subset}{an optional vector specifying a subset of observations to be used in the
fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain
NAs. The default is set by the na.action setting of options, and is na.fail if that
is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL,
no action. Value na.exclude can be useful.}

\item{contrast}{optional lists. See the contrasts.arg of model.matrix.default.}
}
\value{
A fitted model object of class \code{izip} similar to one obtained from \code{glm}
or \code{glm.nb}.

The function \code{summary} (i.e., \code{\link{summary.izip}}) can be used to obtain
and print a summary of the results.

The functions \code{plot} (i.e., \code{\link{plot.izip}}) and
\code{autoplot} can be used to produce a range
of diagnostic plots.

The generic assessor functions \code{coef} (i.e., \code{\link{coef.izip}}),
\code{logLik} (i.e., \code{\link{logLik.izip}})
\code{fitted} (i.e., \code{\link{fitted.izip}}),
\code{nobs} (i.e., \code{\link{nobs.izip}}),
\code{AIC} (i.e., \code{\link{AIC.izip}}) and
\code{residuals} (i.e., \code{\link{residuals.izip}})
can be used to extract various useful features of the value
returned by \code{glm.izip}.

An object class 'glm.izip' is a list containing at least the following components:

\item{coefficients}{a named vector of coefficients}
\item{stderr}{approximate standard errors (using observed rather than expected information) for mean coefficients}
\item{residuals}{the \emph{response} residuals (i.e., observed-fitted)}
\item{fitted_values}{the fitted mean values}
\item{rank}{the numeric rank of the fitted linear model for mean}
\item{linear_predictors}{the linear fit for mean on log scale}
\item{df_residuals}{the residuals degrees of freedom}
\item{df_null}{the residual degrees of freedom for the null model}
\item{null_deviance}{The deviance for the null model.
The null model will include only the intercept.}
\item{deviance; residual_deviance}{The residual deviance of the model}
\item{y}{the \code{y} vector used.}
\item{X}{the model matrix for mean}
\item{model}{the model frame for regression}
\item{call}{the matched call}
\item{formula}{the formula supplied for regression}
\item{terms}{the \code{terms} object used for regression}
\item{data}{the \code{data} argument}
\item{offset}{the \code{offset} vector used}
}
\description{
The function \code{glm.izip} is used to fit an interpretable zero-inflated Poisson
generalized linear model with a log-link.
}
\details{
Fit an interpretable zero-inflated Poisson regression using maximum likelihood estimation.

The model is

\deqn{Y_i ~ ZIP_{\nu}(\mu_i | \lambda = \lambda_{ref}),}

where

\deqn{E(Y_i) = \mu_i = exp(x_i^T \beta),}

\eqn{x_i} are some covariates.

\eqn{\nu \ge 0} is the baseline zero-inflated odds relative to a Poisson with
rate \eqn{\lambda_{ref}}.
}
\examples{
data()
}
\references{
Huang, A. and Fung, T. (2020). Zero-inflated Poisson exponential families, with applications to time-series modelling of counts.
}
