% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izip_distribution.R
\name{iZIP_Distribution}
\alias{iZIP_Distribution}
\alias{dizip}
\alias{pizip}
\alias{qizip}
\alias{rizip}
\title{The interpretable Zero-Inflated Poisson (iZIP) Distribution.}
\usage{
dizip(x, mu, nu, pi, ref.lambda = 1, log.p = FALSE)

pizip(q, mu, nu, pi, lower.tail = TRUE, log.p = FALSE)

qizip(p, mu, nu, pi, lower.tail = TRUE, log.p = FALSE)

rizip(n, mu, nu, pi)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu, nu}{mean and baseline zero inflation odds parameters. Must be strictly positive.}

\item{pi}{an alternative way than nu to parametrized the distribution.
Must be between 0 and 1.}

\item{ref.lambda}{numeric; Baseline Poisson rate. Default is 1.}

\item{log.p}{logical; if \code{TRUE}, probabilities/densities \eqn{p} are returned as
\eqn{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P(X \le x)},
otherwise, \eqn{P(X>x)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations. If \code{length(n)} > 1, the length is taken to
be the number required.}
}
\value{
\code{dizip} gives the density, \code{pizip} gives the distribution function, \code{qizip} gives the quantile function, and \code{rizip} generates random deviates.

Invalid arguments will result in return value \code{NaN}, with a warning.

The length of the results is determined by \code{n} for \code{rizip}, and is the maximum
of the lengths of the numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the length of the results.
Only the first argument of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation for the
interpretable Zero-Inflated Poisson with parameter \code{mu} and \code{nu}
}
\examples{
dizip(0:5, mu = 2, nu = 1.2)
pizip(5, mu=2, nu =1.2)
p <- (1:9)/10
qizip(p, mu = 2, nu = 1)
rizip(10, mu = 2, nu = 4)
}
