% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{iziplrtest}
\alias{iziplrtest}
\title{Likelihood Ratio Test for nested iZIP models}
\usage{
iziplrtest(object1, object2, digits = 3)
}
\arguments{
\item{object1}{an object class 'izip', obtained from a call to \code{glm.izip}}

\item{object2}{an object class 'izip', obtained from a call to \code{glm.izip}}

\item{digits}{numeric; minimum number of significant digits to be used for most numbers.}
}
\description{
Perform a likelihood ratio chi-squared test between nested iZIP models.
The test statistics is calculated as \emph{2*(llik- llik_0)}. The test statistics
has degrees of freedom \emph{r} where \emph{r} is the difference in the number of
parameters between the full and null models.
}
\examples{
data(bioChemists)

## Fit full model
M_bioChem.full <- glm.izip(art ~ ., data = bioChemists)

## Fit null model; without phd
M.bioChem.null <- update(M_bioChem.full, . ~ . - phd)

## Likelihood ratio test for the nested models
iziplrtest(M_bioChem.full, M.bioChem.null) # order of objects is not important

}
\seealso{
\code{\link{glm.izip}}, \code{\link{update.izip}}
}
