% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{predict.izip}
\alias{predict.izip}
\title{Model Predictions for a \code{iZIP} Object}
\usage{
\method{predict}{izip}(
  object,
  newdata = NULL,
  se.fit = FALSE,
  type = c("link", "response"),
  ...
)
}
\arguments{
\item{object}{an object class 'izip', obtained from a call to \code{glm.izip}.}

\item{newdata}{optionally, a data frame in which to look for variables with which to
predict. If omitted, the fitted linear predictors are used.}

\item{se.fit}{logical; indicating if standard errors are required.}

\item{type}{the type of prediction required. The default is 'link' which is the scale
of the linear predictor i.e., a log scale; the alternative 'response' is on the scale
of the response variable. The value of this argument can be abbreviated.}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
If \code{se.fit = FALSE}, a vector of predictions.

If \code{se.fit = TRUE}, a list with components
\item{fit}{Predictions, as for se.fit = FALSE.}
\item{se.fit}{Estimated standard errors.}
}
\description{
This is a function for obtaining predictions and optionally estimates standard
errors of those prediction from a fitted iZIP regression object.
}
\details{
If newdata is omitted the predictions are based on the data used for the fit.
}
\examples{
data(bioChemists)
M_bioChem <- glm.izip(art ~ ., data = bioChemists)

predict(M_bioChem)
predict(M_bioChem, type = "response")
predict(M_bioChem, se.fit = TRUE, type = "response")

newdataframe <- data.frame(
  fem = "Women", mar = "Married", kid5 = 2,
  phd = 3, ment = 8
)
predict(M_bioChem, se.fit = TRUE, newdata = newdataframe, type = "response")
}
