% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{print.izip}
\alias{print.izip}
\title{Print Values of iZIP Model}
\usage{
\method{print}{izip}(x, ...)
}
\arguments{
\item{x}{an object class 'izip', obtained from a call to \code{glm.izip}.}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\description{
\code{print} method for class \code{izip}.
}
\details{
\code{print.izip} can be used to print a short summary of object class 'izip'.
}
\examples{
## For examples see example(glm.izip)
}
\seealso{
\link{summary.izip}, \link{coef.izip}, \link{fitted.izip}, \link{glm.izip}.
}
