% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{print.tsizip}
\alias{print.tsizip}
\title{Print Values of iZIP INGARCH Model}
\usage{
\method{print}{tsizip}(x, ...)
}
\arguments{
\item{x}{an object class 'tsizip', obtained from a call to \code{tsglm.izip}.}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\description{
\code{print} method for class \code{tsizip}.
}
\details{
\code{print.tsizip} can be used to print a short summary of object class 'tsizip'.
}
\examples{
## For examples see example(tsglm.izip)
}
\seealso{
\link{summary.izip}, \link{coef.tsizip}, \link{fitted.tsizip}, \link{tsglm.izip}.
}
