% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{regression.diagnostic.izip}
\alias{regression.diagnostic.izip}
\alias{influence.izip}
\alias{hatvalues.izip}
\alias{rstandard.izip}
\alias{cooks.distance.izip}
\title{iZIP Regression Diagnostic}
\usage{
\method{influence}{izip}(model, ...)

\method{hatvalues}{izip}(model, ...)

\method{rstandard}{izip}(
  model,
  infl = influence.izip(model),
  type = c("deviance", "pearson"),
  ...
)

\method{cooks.distance}{izip}(
  model,
  infl = influence(model),
  res = infl$pear_res,
  hat = infl$hat,
  ...
)
}
\arguments{
\item{model}{an object class 'izip', obtained from a call to \code{\link{glm.izip}}.}

\item{...}{other arguments passed to or from other methods  (currently unused).}

\item{infl}{influence structure as returned by \code{\link{influence}}, only for  \code{\link{rstudent}} and \code{\link{cooks.distance}}.}

\item{type}{type of residuals for \code{\link{rstandard}}. The alternatives are: 'deviance' (default), and 'pearson'.}

\item{res}{residuals, with proper default.}

\item{hat}{hat values \eqn{H[i,i]}, see default.}
}
\description{
This suite of functions provides the basic quantities which are used in forming a wide variety of diagnostics for checking the quality of regression fits.
}
\examples{
data(bioChemists)
M_bioChem <- glm.izip(art~ ., data=bioChemists)
influence(M.attendance)
hatvalues(M.attendance)
rstandard(M.attendance, type = "pearson")
cooks.distance(M.attendance)
}
