% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{residuals.izip}
\alias{residuals.izip}
\title{Extract iZIP Model Residuals}
\usage{
\method{residuals}{izip}(object, type = c("deviance", "pearson", "response"), ...)
}
\arguments{
\item{object}{an object class 'izip', obtained from a call to \code{glm.izip}.}

\item{type}{the \code{type} of residuals which should be returned. The alternatives are:
'deviance' (default), 'pearson' and 'response'. Can be abbreviated.}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
Residuals extracted from the object \code{object}.
}
\description{
\code{residuals} is a generic function which extracts model residuals from objects
returned by the modelling function \code{glm.izip}. \code{resid} is an alias for
\code{residuals} .
}
\seealso{
\code{\link{coef.izip}}, \code{\link{fitted.izip}}, \code{\link{glm.izip}}
}
