% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{summary.tsizip}
\alias{summary.tsizip}
\alias{print.summary.tsizip}
\title{Summarizing a iZIP INGARCH Model Fit}
\usage{
\method{summary}{tsizip}(object, ...)

\method{print}{summary.tsizip}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{an object class 'tsizip', obtained from a call to \code{tsglm.izip}.}

\item{...}{other arguments passed to or from other methods  (currently unused).}

\item{x}{a result of the \emph{default} method of \code{summary()}.}

\item{digits}{numeric; minimum number of significant digits to be used for most numbers.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}
}
\value{
\code{summary.tsizip} returns an object of class "summary.tsizip", a list containing at least the following components:
\item{call}{the component from object.}
\item{family}{the component from object.}
\item{coefficients}{the matrix of coefficients, standard errors, z-values and p-values.}
\item{df}{a 3-vector of the rank of the model and the number of residual degrees of freedom, plus number of coefficients.}
}
\description{
\code{summary} method for class \code{tsizip}.
}
\details{
\code{print.summary.tsizip} tries to be smart about formatting the coefficients, standard errors
and gives 'significance stars'. The \code{coefficients} component of the result gives the
estimated coefficients and their estimated standard errors, together with their ratio. This
third column is labelled as \code{Z value} as the dispersion is fixed for this family. A
forth column gives the two-tailed p-value corresponding to \code{Z value} based on
the asymptotic Normal reference distribution.
}
\examples{
## For examples see example(tsglm.izip)
}
\seealso{
\link{coef.tsizip}, \link{fitted.tsizip}, \link{tsglm.izip}.
}
