% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipingarch.R
\name{tsglm.izip}
\alias{tsglm.izip}
\title{Title}
\usage{
tsglm.izip(
  y,
  X = NULL,
  past_mean_lags = 1,
  past_obs_lags = 1,
  ref.lambda = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
