% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_related.R
\name{tsizip_graphics}
\alias{tsizip_graphics}
\alias{plot.tsizip}
\alias{autoplot.tsizip}
\title{Plot Diagnostic for a \code{tsglm.izip} Object}
\usage{
\method{plot}{tsizip}(
  x,
  which = c(1L, 3L, 5L, 7L, 8L, 9L),
  ask = prod(par("mfcol")) < length(which) && dev.interactive(),
  bins = 10,
  ...
)

\method{autoplot}{tsizip}(
  object,
  which = c(1L, 6L, 7L, 10L, 11L),
  bins = 10,
  ask = TRUE,
  nrow = NULL,
  ncol = NULL,
  output_as_ggplot = TRUE,
  ...
)
}
\arguments{
\item{x}{an object class 'tsizip' object, obtained from a call to \code{tsglm.izip}}

\item{which}{if a subset of plots is required, specify a subset of the numbers 1:11.
See 'Details' below.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot.}

\item{bins}{numeric; the number of bins shown in the PIT histogram or the
PIT Q-Q plot.}

\item{...}{other arguments passed to or from other methods (currently unused).}

\item{object}{an object class 'tsizip' object, obtained from a call to \code{tsglm.izip}}

\item{nrow}{numeric; (optional) number of rows in the plot grid.}

\item{ncol}{numeric; (optional) number of columns in the plot grid.}

\item{output_as_ggplot}{logical; if \code{TRUE}, the function would
return a list of \code{ggplot} objects; if \code{FALSE}, the
function would return an \code{ggarrange} object, which is a ggplot or a list of ggplot.}
}
\value{
A ggarrange object, which is a ggplot or a list of ggplot for autoplot.
}
\description{
Eleven plots (selectable by \code{which}) are currently available using ggplot or graphics:
a plot of observed vs fitted time series plot,
an ACF plot of Pearson residuals,
a PACF plot of Pearson residuals,
a plot of Pearson residuals against time,
a Normal Q-Q Plot of Pearson residuals,
an non-randomized PIT histogram,
an uniform Q-Q plot for non-randomized PIT,
a histogram of normal randomized residuals,
a Q-Q plot of the normal randomized residuals,
an ACF plot of the normal randomized residuals
a PACF plot of the normal randomized residuals
}
\details{
By default, 6 plots (number 1, 6, 7, 8, 10 and 11 from this list of plots) are provided for \code{plot} and 5 plots (number 1, 6, 7, 10 and 11) are provided for \code{autoplot}

There are two plots based on the non-randomized probability integral transformation (PIT)
using \code{\link{izipPIT}}. These are a histogram and a uniform Q-Q plot. If the
model assumption is appropriate, these plots should reflect a sample obtained
from a uniform distribution.

There are also two plots based on the normal randomized residuals calculated
using \code{\link{izipnormRandPIT}}. These are a histogram and a normal Q-Q plot. If the model
assumption is appropriate, these plots should reflect a sample obtained from a normal
distribution.
}
\examples{
data(arson)
M_arson <- tsglm.izip(arson ~ 1, past_mean_lags = 1, past_obs_lags = c(1, 2))
## The default plots are shown
plot(M_arson) # or autoplot(M_arson)

## The plots for the ACF and PACF of the normal randomized residuals
plot(M_arson, which = c(10, 11))
# or autoplot(M_arson, which = c(10,11))
}
\seealso{
\code{\link{izipPIT}}, \code{\link{izipnormRandPIT}},
and \code{\link{tsglm.izip}}.
}
