% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{update.izip}
\alias{update.izip}
\title{Update and Re-fit a iZIP Model}
\usage{
\method{update}{izip}(object, formula., ..., evaluate = TRUE)
}
\arguments{
\item{object}{an object class 'izip', obtained from a call to \code{glm.izip}.}

\item{formula.}{changes to the existing formula in \code{object} -- see \code{update.formula}}

\item{...}{other arguments passed to or from other methods  (currently unused).}

\item{evaluate}{logical; if \code{TRUE} evaluate the new call otherwise simply return
the call}
}
\description{
\code{update} (i.e., \code{update.izip}) will update and (by-default) re-fit a model. It is
identical to \code{update} in the \code{stats} package.
}
\examples{

# To update the regression formula
data(bioChemist)

## Fit full model
M_bioChem.full <- glm.izip(art ~ ., data = bioChemists)
M_bioChem.full

## Dropping phd
M.bioChem.null <- update(M_bioChem.full, . ~ . - phd)
M.bioChem.null

}
\seealso{
\code{\link{glm.izip}}, \code{\link{update.formula}}, \code{\link{iziplrtest}}.
}
