% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{aic.izip}
\alias{aic.izip}
\alias{AIC.izip}
\alias{AIC.tsizip}
\title{Akaike's Information Criterion}
\usage{
\method{AIC}{izip}(object, ..., k = 2)

\method{AIC}{tsizip}(object, ..., k = 2)
}
\arguments{
\item{object}{an object class 'izip' or 'tsizip' object, obtained from a call to \code{glm.izip} or \code{tsglm.izip}.}

\item{...}{other arguments passed to or from other methods  (currently unused).}

\item{k}{numeric: the \emph{penalty} per parameter to be used; the default
k = 2 is the classical AIC.}
}
\value{
A numeric value with the corresponding AIC (or BIC, or ..., depends on k).
}
\description{
A function calculating Akaike's Information Criterion (AIC) based on the log-likelihood
value extracted from \code{\link{logLik.izip}}, according to the formula
\emph{-2\emph{log-likelihood + k}npar}, where \emph{npar} represents the number of parameters
in the fitted model, and \emph{k=2} for the usual AIC or \emph{k=log(n)} (\emph{n} being
the number of observations) for the so-called BIC (Bayesian Information Criterion).
}
\details{
When comparing models fitted by maximum likelihood to the same data, the smaller the AIC or
BIC, the better the fit.
}
\examples{
data(bioChemists)
M_bioChem <- glm.izip(art ~ ., data = bioChemists)
# AIC
AIC(M_bioChem)

data(arson)
M_arson <- tsglm.izip(arson ~ 1, past_mean_lags = 1, past_obs_lags = c(1, 2))
# BIC
AIC(M_arson, k = log(nobs(M_arson)))
}
\seealso{
\link{logLik.izip}, \link{nobs.izip}, \link{glm.izip}, \link{logLik.tsizip}, \link{nobs.tsizip}, \link{tsglm.izip}
}
