% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{fitted.izip}
\alias{fitted.izip}
\title{Extract Fitted Values from an iZIP Model Fit}
\usage{
\method{fitted}{izip}(object, ...)
}
\arguments{
\item{object}{an object class 'izip' object, obtained from a call to \code{glm.izip}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
Fitted values \code{mu} extracted from the object \code{object}.
}
\description{
An accessor function used to extract the fitted values from a 'izip' object.
\code{fitted.values} is an alias for \code{fitted}.
}
\seealso{
\link{coef.izip}, \link{residuals.izip}, \link{glm.izip}.
}
