% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{logLik.izip}
\alias{logLik.izip}
\alias{print.logLik.izip}
\title{Extract the (Maximized) Log-Likelihood from an izip Model Fit}
\usage{
\method{logLik}{izip}(object, ...)

\method{print}{logLik.izip}(x, ...)
}
\arguments{
\item{object}{an object of class 'izip' object, obtained from a call to \code{glm.izip}}

\item{...}{other arguments passed to or from other methods  (currently unused).}

\item{x}{an object of class 'logLik.izip', obtained from a call to \code{logLik.izip}.}
}
\description{
An accessor function used to extract the (maximized) log-likelihood from a 'izip' object.
}
\seealso{
\link{coef.izip}, \link{fitted.izip}, \link{glm.izip}
}
