% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{model.frame.izip}
\alias{model.frame.izip}
\title{Extract the Model Frame from an izip Model Fit}
\usage{
\method{model.frame}{izip}(formula, ...)
}
\arguments{
\item{formula}{an object class 'izip' object, obtained from a call to \code{glm.izip}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
The method will return the saved \code{\link{data.frame}} used when fitting the izip model.
}
\description{
An accessor function used to extract the model frame from a 'izip' object.
}
\seealso{
\link{coef.izip}, \link{residuals.izip}, \link{glm.izip}.
}
