% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{nobs.izip}
\alias{nobs.izip}
\title{Extract the Number of Observation from an iZIP Model Fit}
\usage{
\method{nobs}{izip}(object, ...)
}
\arguments{
\item{object}{an object class 'izip' object, obtained from a call to \code{glm.izip}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
The number of observations extracted from the object \code{object}.
}
\description{
An accessor function used to extract the number of observation from a 'izip' object.
}
\seealso{
\code{\link{coef.izip}}, \code{\link{fitted.izip}}, \code{\link{glm.izip}}
}
