% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{vcov.izip}
\alias{vcov.izip}
\title{Extracting the Variance-Covariance Matrix from a iZIP
Model Fit}
\usage{
\method{vcov}{izip}(object, ...)
}
\arguments{
\item{object}{an object class 'izip' object, obtained from a call to \code{glm.izip}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
The method will return the estimated covariances between the parameter estimates of the fitted iZIP model.
}
\description{
Extracting the Variance-Covariance Matrix from a iZIP
Model Fit
}
\examples{
data(bioChemists)
M_bioChem <- glm.izip(art~ ., data=bioChemists)
vcov(M_bioChem)
}
